// types.ts
export interface FormData {
  presentadoPara: string;
  sesionDia: string;
  sesionMes:  string;
  sesionAnio: string;
  facilitador: string;
  nombreCommit: string;
  futuroImparable: string[];
  queVendes: {
    principal: string;
    notas: string;
  };
  clienteIdeal: string;
  leadMagnet: {
    tipo: string;
    regalo: string;
    titulo: string;
    cta: string;
  };
  optInPage: {
    titulo: string;
    subtitulo: string;
    beneficio1: string;
    beneficio2: string;
    beneficio3: string;
  };
  emailSwell: {
    bienvenida: string;
    prueba: string;
    invitacion: string;
  };
  sistemaEntrega: string;
  declaracionSolucion: {
    quien: string;
    que: string;
    como: string;
    resultado: string;
  };
}